# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API LIVE TEST",
        "description": "TLYN API V1",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "دریافت لیست سفارشات باز",
                "description": "دریافت لیست سفارشات باز",
                "operationId": "00d3856d2d3d93124e0cc021d0cb5b79",
                "responses": {
                    "200": {
                        "description": "لیست سفارشات باز",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "عدم احراز هویت"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "ثبت سفارش جدید",
                "description": "ثبت سفارش جدید",
                "operationId": "79da0a1adcc617490783dfd61927ff56",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "price",
                                    "weight"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "buy",
                                            "sell"
                                        ],
                                        "example": "buy"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10000000
                                    },
                                    "weight": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "سفارش با موفقیت ثبت شد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Order successfully placed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "خطا در ثبت سفارش",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "There is not enough cash to place an order."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "عدم احراز هویت"
                    },
                    "422": {
                        "description": "خطای اعتبارسنجی",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "price": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The price field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/orders/{id}/cancel": {
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "لغو سفارش",
                "description": "لغو سفارش",
                "operationId": "6adc37bd99c10a829acd309b0aa27ba1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "شناسه سفارش",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "سفارش با موفقیت لغو شد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Order canceled"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "خطا در لغو سفارش",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Only open orders can be canceled"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "عدم احراز هویت"
                    },
                    "404": {
                        "description": "سفارش یافت نشد"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/orders/history": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "تاریخچه سفارشات کاربر",
                "description": "تاریخچه سفارشات کاربر",
                "operationId": "dfd51cb81ec003c6aedd5796542f4854",
                "responses": {
                    "200": {
                        "description": "لیست تاریخچه سفارشات",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "عدم احراز هویت"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Order": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "buy",
                            "sell"
                        ],
                        "example": "buy"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 10000000
                    },
                    "weight": {
                        "type": "number",
                        "format": "float",
                        "example": 2.5
                    },
                    "remaining_weight": {
                        "type": "number",
                        "format": "float",
                        "example": 2.5
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "open",
                            "filled",
                            "canceled"
                        ],
                        "example": "open"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": " Use login to get the passport token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "Passport",
                "scheme": "bearer"
            }
        }
    },
...
}